requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
pandas>=2.0.0
dashscope>=1.14.0
pyperclip>=1.8.2  # 新增：复制剪贴板依赖